Assignment-1/Task-1/requirements.txt
numpy==1.14.2
Pillow==5.4.1
